BREAKING CHANGE: re-organize project structure
- /ExchangeDiary_Server/exchange-diary -> /exchange-diary
