+++ application/route/diaryRouter.go
@@ -1,0 +1,7 @@
+package route
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+func DiaryRoutes(incomingRoutes *gin.Engine) {}
