--- exchange-diary/application/route/diaryRouter.go
@@ -1,7 +1,0 @@
-package route
-
-import (
-	"github.com/gin-gonic/gin"
-)
-
-func DiaryRoutes(incomingRoutes *gin.Engine) {}
