+++ go.mod
@@ -1,0 +1,20 @@
+module github.com/exchange-diary
+
+go 1.15
+
+require (
+	github.com/gin-contrib/zap v0.0.2
+	github.com/gin-gonic/gin v1.7.7
+	github.com/go-playground/validator/v10 v10.10.0 // indirect
+	github.com/jinzhu/copier v0.3.5
+	github.com/spf13/viper v1.10.1
+	github.com/ugorji/go v1.2.6 // indirect
+	go.uber.org/atomic v1.9.0 // indirect
+	go.uber.org/multierr v1.7.0 // indirect
+	go.uber.org/zap v1.21.0
+	golang.org/x/crypto v0.0.0-20220214200702-86341886e292 // indirect
+	golang.org/x/sys v0.0.0-20220209214540-3681064d5158 // indirect
+	golang.org/x/tools v0.1.9 // indirect
+	gorm.io/driver/mysql v1.3.2
+	gorm.io/gorm v1.23.1
+)
