+++ infrastructure/configs/dev.yaml
@@ -1,0 +1,12 @@
+db-config:
+  host: "0.0.0.0"
+  port: 3306
+  user: "root"
+  name: "voda"
+  password: "<로컬_DB_비밀번호>"
+
+kakao-client:
+  oauth:
+    client-id: "c663ce96b79395b5dcc6d7f774e1af67"
+    client-secret: "LLlLMwzJ9VjLXhjsJYx4pE3zsE1x9m4P"
+    redirect-url: "http://localhost:8080/auth/kakao"
